FIX #48: Add offset to resize Zoom. CAUTION, read below
Artscript creates the perfect (or almost) in zoom mode,
setting both x and y offset values with zoom to any corner
is not recommended as one (x or y, depends on image) will
make the image move beyond the canvas. You will get a
transparent stripe.

When setting offset values and a corner placement be sure
to know if vertical or horizontal image data is available