FIX #59: Calligraconvert hangs artscript if run with no kde daemon
This checks if kdeinit4 is running, if it isn't run it. This aboids
a bug that made artscript hang when calligra called kdeinit4 as it's
startup process since it skipped a TERM signal making artscript wait
forever.

This also fix GIMP ora handling bugs as we go back to Calligra to
handle ORA as it does it properly.

NOTICE: Sometimes calligraconvert crashes on rare ocassions.